bottomless: add read-only replicas that read from S3
This commit adds support for using bottomless in "read-only replica"
mode. In this mode, all data is read exclusively from S3
and nothing gets restored to disk.

FIXME: it also contains a terrible, terrible hack in order
to work - on startup, the first page of the database is restored
from S3, and the database file is truncated to match the original
DB size, thus making libSQL think it's a properly structured
database file. After that, all data is read only from S3.

A proper solution is to ship bottomless WAL with a bottomless VFS
implementation - with the VFS set up to read pages from S3.
Effectively, it will only ever read the first db page from there,
and from then on it will just read everything from S3.