sqld-libsql-bindings: deregister methods before freeing them
If the WAL methods are not deregistered from a global libSQL
table before dropping them, we end up with having a dangling
pointer in the libSQL structure. It often worked, because
the freed memory just didn't manage to get overwritten yet,
but it was a time bomb nonetheless.

The scenario was as follows:
1. Fiber A registers WAL methods W1
2. Fiber A frees WAL methods W1, but does not deregister them
   from the libSQL global data structure
3. Fiber B registers WAL methods W2, and registration involves
   iterating over the data structure that contains a dangling
   pointer, after the methods were freed by (2.)