treewide: integrate bottomless as WAL hook
From now on, bottomless can be enabled in sqld as a regular WalHook.
TODO: the main functionality, i.e. actually replicating frames is not hooked to
anything yet. It will be done, but it should be heavily considered to just
queue writes with an unbounded channel (which allows sync insertions), so that
everything happens in the background.