replication/primary: unconditionally update max frame number
With gRPC replication, it was reasonable to assume that there
are listeners to the max frame number notifier, but with HTTP
it's not necessarily the case. Since watch::send() fails if there are
no receivers, we hereby switch to send_replace(), which successfully
updates the value even if there are no active receivers.