hrana: add diagnostics for connections
This commit adds a /v2/diagnostics endpoint which prints
various information about current hrana-over-http connections.

Draft, because the diagnostics are currently in a very debuggy
format, and I'm figuring out if we can make it more human-readable.
Still, they're enough to determine if something is holding a lock
via an abandoned hrana-over-http stream.

Example:
```
$ curl -s http://localhost:8080/v2/diagnostics | jq
[
  "expired",
  "expired",
  "expired",
  "expired",
  "expired",
  "(conn: Mutex { data: <locked> }, timeout_ms: 872, stolen: false)",
  "(conn: Mutex { data: <locked> }, timeout_ms: 0, stolen: true)"
]
```