bottomless: remove false positive [BUG] message
For a transaction that spans multiple xFrames calls, it's ok
if our last_valid_frame doesn't match the one that came with xFrames.
As long as our last_valid_frame is *not greater* than the one reported,
we're good. If it's greater, we're in trouble and we report a bug.