connection: vacuum before checkpointing
This commit adds an optional VACUUM operation before our
periodic checkpoint -- since VACUUM should always be followe
by a checkpoint.
The are two criteria to qualify for vacuuming:
1. We have at least 32MiB of data
2. There are more free pages in the db than regular ones

If that's the case, we keep the db file more than 2x larger
than it could be, and that calls for a vacuum.
Also, for small databases, we entirely skip vacuuming.

Fixes #734